+++ package.json
@@ -10,6 +10,7 @@
     "mockman-js": "^1.0.5",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.2.2",
     "react-scripts": "5.0.0",
     "uuid": "^8.3.2"
   },
