--- src/components/Homepage.js
@@ -1,16 +1,0 @@
-import '../css/home-page.css';
-import { HomePageCategories, HomePageCards, HomePageHero } from '../components/Homepage/';
-
-
-const HomePage = () => {
-
-    return <>
-      <HomePageCategories />
-      <HomePageHero />
-      <HomePageCards />
-    </>
-    
-    
-}
-
-export { HomePage }
