+++ src/components/Navigation.js
@@ -1,10 +1,9 @@
 import '../css/nav.css';
+import { Link } from 'react-router-dom';
 
 const Navigation = () => {
     return  <nav class="nav-container display-flex primary-clr p-rl-6">
-        <a href="./index.html" class="nav-link font-size-2pt6">
-            Hotelaza
-        </a>
+        <Link to='/' class="nav-link font-size-2pt6">Hotelaza</Link>
         <div class="input-search pos-rltv m-rght-auto">
             <i class="fa-solid fa-magnifying-glass pos-absolute font-size-2pt6"></i>
             <input type="text" placeholder="search" class="nav-input-search"/>
